+++ concordium-consensus/src/Concordium/GlobalState/Persistent/Bakers.hs
@@ -13,6 +13,7 @@ module Concordium.GlobalState.Persistent.Bakers where
 
 import Control.Exception
 import Control.Monad
+import Control.Monad.Trans
 import qualified Data.Map.Strict as Map
 import qualified Data.Set as Set
 import qualified Data.Vector as Vec
@@ -23,7 +24,9 @@ import Data.Foldable (foldlM)
 import Concordium.GlobalState.BakerInfo
 import qualified Concordium.GlobalState.Basic.BlockState.Bakers as Basic
 import qualified Concordium.Types.Accounts as BaseAccounts
+import Concordium.GlobalState.Parameters
 import Concordium.GlobalState.Persistent.Account
+import qualified Concordium.GlobalState.Persistent.Accounts as Accounts
 import Concordium.GlobalState.Persistent.BlobStore
 import Concordium.Types
 import Concordium.Types.Execution (DelegationTarget(..))
@@ -41,6 +44,14 @@ import Concordium.Utils.Serialization.Put
 newtype BakerInfos (av :: AccountVersion)
     = BakerInfos (Vec.Vector (PersistentBakerInfoEx av))
     deriving (Show)
+
+-- |See documentation of @migratePersistentBlockState@.
+migrateBakerInfos ::
+    forall oldpv pv t m.
+    ( IsProtocolVersion pv
+    , SupportMigration m t
+    ) => StateMigrationParameters oldpv pv -> BakerInfos (AccountVersionFor oldpv) -> t m (BakerInfos (AccountVersionFor pv))
+migrateBakerInfos migration (BakerInfos inner) = BakerInfos <$> mapM (migratePersistentBakerInfoEx migration) inner
 
 instance (IsAccountVersion av, MonadBlobStore m) => BlobStorable m (BakerInfos av) where
     storeUpdate (BakerInfos v) = do
@@ -92,6 +103,39 @@ data PersistentEpochBakers (av :: AccountVersion) = PersistentEpochBakers {
 } deriving (Show)
 
 makeLenses ''PersistentEpochBakers
+
+-- |Extract the list of pairs of (baker id, staked amount). The list is ordered
+-- by increasing 'BakerId'.
+-- The intention is that the list will be consumed immediately.
+extractBakerStakes :: (IsAccountVersion av, MonadBlobStore m) => PersistentEpochBakers av -> m [(BakerId, Amount)]
+extractBakerStakes PersistentEpochBakers{..} = do
+  BakerInfos infos <- refLoad _bakerInfos
+  BakerStakes stakes <- refLoad _bakerStakes
+  zipWithM (\bi bs -> do
+               bid <- loadBakerId bi
+               return (bid, bs)
+               )
+      (Vec.toList infos)
+      (Vec.toList stakes)
+
+-- |See documentation of @migratePersistentBlockState@.
+migratePersistentEpochBakers ::
+    forall oldpv pv t m.
+    ( IsProtocolVersion oldpv
+    , IsProtocolVersion pv
+    , SupportMigration m t
+    ) =>
+    StateMigrationParameters oldpv pv ->
+    PersistentEpochBakers (AccountVersionFor oldpv) ->
+    t m (PersistentEpochBakers (AccountVersionFor pv))
+migratePersistentEpochBakers migration PersistentEpochBakers {..} = do
+  newBakerInfos <- migrateHashedBufferedRef (migrateBakerInfos migration) _bakerInfos
+  newBakerStakes <- migrateHashedBufferedRefKeepHash _bakerStakes
+  return PersistentEpochBakers {
+    _bakerInfos = newBakerInfos,
+    _bakerStakes = newBakerStakes,
+    ..
+    }
 
 -- |Look up a baker and its stake in a 'PersistentEpochBakers'.
 epochBaker :: forall m av. (IsAccountVersion av, MonadBlobStore m) => BakerId -> PersistentEpochBakers av -> m (Maybe (BaseAccounts.BakerInfo, Amount))
@@ -197,6 +241,29 @@ data PersistentActiveDelegators (av :: AccountVersion) where
         } ->
         PersistentActiveDelegators 'AccountV1
 
+-- |See documentation of @migratePersistentBlockState@.
+migratePersistentActiveDelegators ::
+    (BlobStorable m (), BlobStorable (t m) (), MonadTrans t) =>
+    StateMigrationParameters oldpv pv ->
+    PersistentActiveDelegators (AccountVersionFor oldpv) ->
+    t m (PersistentActiveDelegators (AccountVersionFor pv))
+migratePersistentActiveDelegators StateMigrationParametersTrivial = \case
+    PersistentActiveDelegatorsV0 -> return PersistentActiveDelegatorsV0
+    PersistentActiveDelegatorsV1{..} -> do
+        newDelegators <- Trie.migrateTrieN True return adDelegators
+        return PersistentActiveDelegatorsV1{adDelegators = newDelegators, ..}
+migratePersistentActiveDelegators StateMigrationParametersP1P2 = \case
+    PersistentActiveDelegatorsV0 -> return PersistentActiveDelegatorsV0
+migratePersistentActiveDelegators StateMigrationParametersP2P3 = \case
+    PersistentActiveDelegatorsV0 -> return PersistentActiveDelegatorsV0
+migratePersistentActiveDelegators (StateMigrationParametersP3ToP4 _) = \case
+    PersistentActiveDelegatorsV0 ->
+        return
+            PersistentActiveDelegatorsV1
+                { adDelegators = Trie.empty
+                , adDelegatorTotalCapital = 0
+                }
+
 emptyPersistentActiveDelegators :: forall av. IsAccountVersion av => PersistentActiveDelegators av
 emptyPersistentActiveDelegators =
     case accountVersion @av of
@@ -232,6 +299,18 @@ data TotalActiveCapital (av :: AccountVersion) where
 
 deriving instance Show (TotalActiveCapital av)
 
+-- |See documentation of @migratePersistentBlockState@.
+migrateTotalActiveCapital ::
+  StateMigrationParameters oldpv pv ->
+  -- |The total amount staked by all the __bakers__.
+  Amount ->
+  TotalActiveCapital (AccountVersionFor oldpv) ->
+  TotalActiveCapital (AccountVersionFor pv)
+migrateTotalActiveCapital StateMigrationParametersTrivial _ x = x
+migrateTotalActiveCapital StateMigrationParametersP1P2 _ x = x
+migrateTotalActiveCapital StateMigrationParametersP2P3 _ x = x
+migrateTotalActiveCapital (StateMigrationParametersP3ToP4 _) bts TotalActiveCapitalV0 = TotalActiveCapitalV1 bts
+
 instance IsAccountVersion av => Serialize (TotalActiveCapital av) where
     put TotalActiveCapitalV0 = return ()
     put (TotalActiveCapitalV1 amt) = put amt
@@ -264,6 +343,39 @@ data PersistentActiveBakers (av :: AccountVersion) = PersistentActiveBakers {
 } deriving (Show)
 
 makeLenses ''PersistentActiveBakers
+
+-- |See documentation of @migratePersistentBlockState@.
+migratePersistentActiveBakers :: forall oldpv pv t m .
+    (IsProtocolVersion oldpv,
+     IsProtocolVersion pv,
+     SupportMigration m t,
+     Accounts.SupportsPersistentAccount pv (t m)
+    ) => 
+    StateMigrationParameters oldpv pv ->
+    -- |Already migrated accounts.
+    Accounts.Accounts pv -> 
+    PersistentActiveBakers (AccountVersionFor oldpv) ->
+    t m (PersistentActiveBakers (AccountVersionFor pv))
+migratePersistentActiveBakers migration accounts PersistentActiveBakers{..} = do
+  newActiveBakers <- Trie.migrateTrieN True (migratePersistentActiveDelegators migration) _activeBakers
+  newAggregationKeys <- Trie.migrateTrieN True return _aggregationKeys
+  newPassiveDelegators <- migratePersistentActiveDelegators migration _passiveDelegators
+  bakerIds <- Trie.keysAsc newActiveBakers
+  totalStakedAmount <- foldM (\acc (BakerId aid) ->
+    Accounts.indexedAccount aid accounts >>= \case
+      Nothing -> error "Baker account does not exist."
+      Just pa -> case pa ^. accountBaker of
+        Null -> error "Baker account not a baker."
+        Some bref -> do
+            bkr <- refLoad bref
+            return $! (acc + _stakedAmount bkr)) 0 bakerIds
+  let newTotalActiveCapital = migrateTotalActiveCapital migration totalStakedAmount _totalActiveCapital
+  return PersistentActiveBakers {
+    _activeBakers = newActiveBakers,
+    _aggregationKeys = newAggregationKeys,
+    _passiveDelegators = newPassiveDelegators,
+    _totalActiveCapital = newTotalActiveCapital
+    }
 
 totalActiveCapitalV1 :: Lens' (PersistentActiveBakers 'AccountV1) Amount
 totalActiveCapitalV1 = totalActiveCapital . tac